add: prenotazioni

+++ components/modal.tsx
@@ -3,8 +3,10 @@ import Router from "next/router"
 import { useEffect, useImperativeHandle } from "react"
 
 const NONE_VAL = "none"
+const ADD = "ADD"
+const DELETE = "DELETE"
 
-function Modal({seatName, yourSeat, username, reserveData, setReserveData, fromTo}: any) { 
+function Modal({seatName, action, username, reserveData, setReserveData, fromTo}: any) { 
 
   useEffect(() => {
     var modal = document.getElementById("myModal") as HTMLElement
@@ -32,7 +34,7 @@ function Modal({seatName, yourSeat, username, reserveData, setReserveData, fromT
     console.log("FROM: " + fromTo.from)
     console.log("TO: " + fromTo.to)
 
-    if(!yourSeat) {
+    if (action === ADD) {
       await axios.post("/api/addReserve", {
         seatId: seatId,
         userId: userId,
@@ -40,7 +42,7 @@ function Modal({seatName, yourSeat, username, reserveData, setReserveData, fromT
         from: new Date(fromTo.from),
         to: new Date(fromTo.to)
       })
-    } else {
+    } else if (action === DELETE) {
       const reserveToDelete = reserveData.find((reserve: any) => reserve.seat.name === seatName)
       await axios.delete("/api/reserve/" + reserveToDelete.id)
     }
@@ -53,9 +55,9 @@ function Modal({seatName, yourSeat, username, reserveData, setReserveData, fromT
     <div className="modal-content">
       <span className="close">&times;</span>
       <div className="modal-body">
-        <p>{!yourSeat ? 
-          "Vuoi procedere con la prenotazione del posto " + seatName + "?" :
-          "Vuoi annullare la prenotazione del posto " + seatName + "?"
+        <p>{action === ADD ? 
+            "Vuoi procedere con la prenotazione del posto " + seatName + "?" : 
+            "Vuoi annullare la prenotazione del posto " + seatName + "?"
           }</p>
         {username === "admin" && <input></input>}
         <button className="modal-button" onClick={() => handleSeat()} >Conferma</button>
